feat: fucking push everything

+++ app/src/main/java/com/example/coinly/db/Transaction.java
@@ -61,7 +61,7 @@ public class Transaction implements Database.MapParser<Transaction> {
         return this;
     }
 
-    public Transaction withAmount(float amount) {
+    public Transaction withAmount(double amount) {
         this.amount = amount;
         return this;
     }
